+++ common/src/main/java/dev/xylonity/bonsai/clockwork/registry/ClockworkPersistentSounds.java
@@ -1,6 +1,7 @@
 package dev.xylonity.bonsai.clockwork.registry;
 
 import dev.xylonity.bonsai.clockwork.common.entity.passive.DragonflyEntity;
+import dev.xylonity.bonsai.clockwork.common.entity.tool.ClockworkDrillEntity;
 import dev.xylonity.knightlib.api.sound.persistent.KnightLibPersistentSounds;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.player.Player;
@@ -9,23 +10,41 @@ public class ClockworkPersistentSounds {
 
     public static void init() {
         KnightLibPersistentSounds.profile(DragonflyEntity.class)
-                .sound("fly")
+                .sound("clockwork:fly")
                     .event(ClockworkSounds.DRAGONFLY_FLY)
-                    .volume(0.4f)
+                    .volume(0.3f)
                     .build()
-                .sound("idle")
+                .sound("clockwork:idle")
                     .event(ClockworkSounds.DRAGONFLY_IDLE)
-                    .volume(0.4f)
+                    .volume(0.3f)
                     .build()
                 .submit();
 
         KnightLibPersistentSounds.profile(Player.class)
-                .sound("spray")
+                .sound("clockwork:spray")
                     .event(ClockworkSounds.SPRAYER_LOOP)
                     .source(SoundSource.PLAYERS)
                     .volume(0.5f)
                     .onStart(ClockworkSounds.SPRAYER_START, 0.6f)
                     .onStop(ClockworkSounds.SPRAYER_END, 0.6f)
+                    .build()
+                .sound("clockwork:flamethrower")
+                    .event(ClockworkSounds.FLAMETHROWER_LOOP)
+                    .source(SoundSource.PLAYERS)
+                    .volume(10f)
+                    .onStart(ClockworkSounds.SPRAYER_START, 0.35f)
+                    .onStop(ClockworkSounds.SPRAYER_END, 0.35f)
+                    .build()
+                .submit();
+
+        KnightLibPersistentSounds.profile(ClockworkDrillEntity.class)
+                .sound("clockwork:drill_loop")
+                    .event(ClockworkSounds.DRILL_LOOP)
+                    .volume(0.5f)
+                    .build()
+                .sound("clockwork:drill_walk_loop")
+                    .event(ClockworkSounds.DRILL_WALK_LOOP)
+                    .volume(0.5f)
                     .build()
                 .submit();